variables,  syntactic Sugar

+++ week1/addition.c
@@ -4,10 +4,10 @@
 
 int main(void)
 {
-    int x = get_int("x: ");
+    long x = get_int("x: ");
 
-    int y = get_int("y: ");
+    long y = get_int("y: ");
 
 
-    printf("%i\n", x+y);
+    printf("%li\n", x+y);
 }
